test(core): split @scope coverage into dedicated suite
- Move @scope tests from resolveStyleRules.test.ts to
  resolveStyleRules.scope.test.ts and add a case for the nested-@scope
  warning.
- Pull the @scope expectations in getStyleBucketName.test.ts into a
  separate it() block noting that @scope does not affect bucket
  strategy.
- Add a hashPropertyKey collision test that pins the scope prefix
  against supports and against an empty scope.
